패키지, 엔티티 수
- 각 탭의 메뉴를 패키지별로 분리.
- 명확한 양방향 관계 확립을 위해 부모 엔티티에 자식 컬렉션 추가
