--- growith/src/main/java/com/example/growith/InqueryType.java
@@ -1,17 +1,0 @@
-package com.example.growith;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import lombok.Data;
-
-@Data
-@Entity
-public class InqueryType {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
-
-    private String content;
-}
