--- growith/src/main/java/com/example/growith/Member.java
@@ -1,26 +1,0 @@
-package com.example.growith;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import lombok.Data;
-
-//관리자 entity
-@Data
-@Entity
-public class Member {
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer id;
-	
-	@Column(unique = true)
-	private String username;
-	
-	private String password;
-	
-	private String email;
-	
-	private String role;
-}
