+++ growith/src/main/java/com/example/growith/customerservice/PhoneNumber.java
@@ -1,4 +1,4 @@
-package com.example.growith;
+package com.example.growith.customerservice;
 
 import jakarta.validation.Constraint;
 import jakarta.validation.Payload;
