+++ growith/src/main/java/com/example/growith/customerservice/PhoneNumberValidator.java
@@ -1,4 +1,4 @@
-package com.example.growith;
+package com.example.growith.customerservice;
 
 import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
