+++ growith/src/main/java/com/example/growith/investor/InvestorController.java
@@ -1,4 +1,4 @@
-package com.example.growith;
+package com.example.growith.investor;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
